+++ ego/src/solver/mod.rs
@@ -1,4 +1,4 @@
-pub mod egor_impl;
 pub mod egor_service;
 pub mod egor_solver;
 pub mod egor_state;
+pub mod solver_impl;
