+++ ego/src/solver/solver_impl.rs
